+++ internal/parser/zz_ast_test.go
@@ -178,6 +178,21 @@ func TestInverseStmt_Pos(t *testing.T) {
 }
 
 // ////////////////////////////////////
+// FLASH
+// ////////////////////////////////////
+func TestFlashStmt_Pos(t *testing.T) {
+	h := &FlashStmt{
+		Line:   12,
+		Column: 3,
+	}
+
+	line, col, tok := h.Pos()
+	testutils.Equal(t, "line", line, 12)
+	testutils.Equal(t, "column", col, 3)
+	testutils.Equal(t, "token", tok, "FLASH")
+}
+
+// ////////////////////////////////////
 // HOME / HTAB / VTAB
 // ////////////////////////////////////
 func TestHomeStmt_Pos(t *testing.T) {