+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a Changelog (https://keepachangelog.com), and this project adheres to Semantic Versioning (https://semver.org).
 
+## [Unreleased] - 2026-01-30
+### Added
+- Add `FLASH` support in Apple II Basic. Add relevant unit tests.
+
 ## [Unreleased] - 2026-01-29
 ### Added
 - Add `NORMAL` support in Apple II Basic. Add relevant unit tests.